rename uthmanic_hafs to text_qpc_hafs We're going to add other qirats from Quran Priniting Complex( QPC ), renaming will make schema consistance. text_qpc_qaloon, text_qpc_warsh etc

+++ db/migrate/20210313031122_create_audio_chapter_audio_files.rb
@@ -6,7 +6,7 @@ class CreateAudioChapterAudioFiles < ActiveRecord::Migration[6.1]
       t.integer :total_files
       t.integer :stream_count
       t.integer :download_count
-      t.integer :file_size
+      t.float :file_size
       t.integer :bit_rate
       t.integer :duration
